Delete all build steps and use raw ESM
