+++ .gitignore
@@ -1,7 +1,2 @@
 node_modules
-
 .DS_Store
-
-package-lock.json
-
-.eslintrc
