+++ README.md
@@ -1,42 +1,9 @@
-# Multipage Coffee Shop Site (React.js)
+# Tyler's Cafe
 
-A front-end multipage web app of a mock coffee shop using front-end routing. Built with React.js.
+A simple React site that displays a menu for my home cafe.
 
-Inspired by the [Restaurant Page assignment](https://www.theodinproject.com/lessons/node-path-javascript-restaurant-page) as part of the curriculum for the [JavaScript course](https://www.theodinproject.com/paths/full-stack-javascript/courses/javascript) on [The Odin Project](https://www.theodinproject.com).
+Forked from [Autumn Bullard's](https://autumnchris.github.io/portfolio) amazing [coffee shop example](https://github.com/autumnchris/multipage-coffee-shop-site-reactjs)!
 
----
+## Getting Started
 
-## Built With
-* [React.js](https://reactjs.org)
-* [React Router](https://reactrouter.com)
-* [Sass](http://sass-lang.com)
-* [Webpack](https://webpack.js.org)
-* [Babel](https://babeljs.io)
-* [Normalize.css](https://necolas.github.io/normalize.css)
-* [Font Awesome](https://fontawesome.com)
-* [Google Fonts](https://fonts.google.com)
-* [Dark Idea font](https://befonts.com/dark-idea-font.html)
-* [TA Typefire font](https://befonts.com/ta-typefire-font-family.html)
-
-# Demo
-
-View project demo at [https://autumnchris-multipage-coffee-shop-site.onrender.com](https://autumnchris-multipage-coffee-shop-site.onrender.com).
-
-## Instructions
-
-After forking and cloning, navigate to the repository in your command line and install the NPM packages:
-```
-npm install
-```
-
-Run the following script in your command line to run the application:
-```
-npm start
-```
-
-Once the server is running, go to `http://localhost:8080` in your browser.
-
-Before committing any changes, run the following script to update your static files for production:
-```
-npm run build
-```
+This project doesn't have any build steps. It leverages ESM & HTM to render the React app on the fly.
