+++ package.json
@@ -1,40 +1,20 @@
 {
-  "name": "multipage-coffee-shop-site-reactjs",
+  "name": "tylers-cafe",
   "version": "1.0.0",
-  "description": "A front-end multipage web app of a mock coffee shop using front-end routing. Built with React.js.",
+  "description": "A simple React site that displays a menu for my home cafe.",
+  "private": true,
   "main": "index.js",
   "scripts": {
-    "build": "webpack --mode production",
-    "start": "webpack serve --mode development"
+    "start": "npx serve public/"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/autumnchris/multipage-coffee-shop-site-reactjs.git"
+    "url": "git+https://github.com/TylerLeonhardt/home-coffee-shop-site.git"
   },
-  "author": "autumnchris",
-  "license": "ISC",
+  "author": "TylerLeonhardt",
+  "license": "MIT",
   "bugs": {
-    "url": "https://github.com/autumnchris/multipage-coffee-shop-site-reactjs/issues"
+    "url": "https://github.com/TylerLeonhardt/home-coffee-shop-site/issues"
   },
-  "homepage": "https://autumnchris-multipage-coffee-shop-site.onrender.com",
-  "dependencies": {
-    "normalize.css": "^5.0.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-router-dom": "^6.4.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.19.1",
-    "@babel/preset-env": "^7.19.1",
-    "@babel/preset-react": "^7.18.6",
-    "babel-loader": "^8.2.5",
-    "css-loader": "^6.7.1",
-    "file-loader": "^6.2.0",
-    "mini-css-extract-plugin": "^2.6.1",
-    "sass": "^1.54.9",
-    "sass-loader": "^13.0.2",
-    "webpack": "^5.74.0",
-    "webpack-cli": "^4.10.0",
-    "webpack-dev-server": "^4.11.0"
-  }
+  "homepage": "https://tylers.cafe"
 }
