+++ public/index.html
@@ -5,16 +5,13 @@
     <title>☕️ by Tyler</title>
     <link rel="icon" href="favicon.ico">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.15.4/css/all.css">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Playfair+Display&family=Source+Sans+Pro:wght@400;600&display=swap">
-    <link rel="stylesheet" href="style.css">
+    <link href="https://use.fontawesome.com/releases/v5.15.4/css/all.css" rel="stylesheet">
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Meow+Script&family=Playfair+Display:wght@400;600&display=swap" rel="stylesheet">
+    <link href="style.css" rel="stylesheet">
+    <script type="module" src="./index.js"></script>
   </head>
-
   <body>
-    <div id="app"></div>
-    <script src="index.bundle.js"></script>
-    <noscript>
-      <p class="message info-message"><span class="fa fa-info-circle fa-lg fa-fw" aria-hidden="true"></span> This site uses JavaScript. Please enable JavaScript in your browser.</p>
-    </noscript>
   </body>
 </html>
