--- src/data/image-gallery.js
@@ -1,64 +1,0 @@
-const imageGallery = [
-  {
-    id: 1,
-    path: require('../assets/images/coffee-grounds.jpg'),
-    alt: 'photo of scoop in a bowl of coffee grounds'
-  },
-    {
-    id: 2,
-    path: require('../assets/images/barista-pouring-cream.jpg'),
-    alt: 'photo of coffee cup held by a barista who is pouring cream into it off camera'
-  },
-  {
-    id: 3,
-    path: require('../assets/images/avocado-toast.jpg'),
-    alt: 'photo of avocado toast on a plate topped with everything bagel seasoning'
-  },
-  {
-    id: 4,
-    path: require('../assets/images/frappe.jpg'),
-    alt: 'photo of frappe on a table in a mason jar with a straw'
-  },
-  {
-    id: 5,
-    path: require('../assets/images/croissant.jpg'),
-    alt: 'photo of fresh baked croissants'
-  },
-  {
-    id: 6,
-    path: require('../assets/images/wine-bottles.jpg'),
-    alt: 'photo of a rack of wine bottles'
-  },
-  {
-    id: 7,
-    path: require('../assets/images/coffee-beans.jpg'),
-    alt: 'photo of coffee beans'
-  },
-  {
-    id: 8,
-    path: require('../assets/images/matcha-latte.jpg'),
-    alt: 'photo of a matcha latte'
-  },
-  {
-    id: 9,
-    path: require('../assets/images/laptop.jpg'),
-    alt: 'photo of an open laptop sitting on a table in a cafe next to a cup of iced coffee'
-  },
-  {
-    id: 10,
-    path: require('../assets/images/wine-glasses.jpg'),
-    alt: 'photo of two people off camera toasting with two glasses of red wine'
-  },
-  {
-    id: 11,
-    path: require('../assets/images/latte.jpg'),
-    alt: 'photo of barista off camera pouring cream into a latte'
-  },
-  {
-    id: 12,
-    path: require('../assets/images/espresso-grounds.jpg'),
-    alt: 'photo of a scoop of espresso grounds'
-  }
-];
-
-export default imageGallery;
