--- src/data/menu.js
@@ -1,76 +1,0 @@
-const menu = [
-  {
-    id: 10,
-    itemName: 'Pour over ✧',
-    price: 'hot ✧ or iced - 8oz / 12oz',
-    type: 'Coffee'
-  },
-  {
-    id: 11,
-    itemName: 'French press',
-    price: 'hot or iced - 8oz / 12oz',
-    type: 'Coffee'
-  },
-  {
-    id: 21,
-    itemName: 'Latte ✧',
-    price: 'hot ✧ or iced - 6oz ✧ / 8oz / 12oz',
-    type: 'Espresso'
-  },
-  {
-    id: 27,
-    itemName: 'Shaken espresso ✧',
-    price: '8oz',
-    type: 'Espresso',
-  },
-  {
-    id: 22,
-    itemName: 'Cappuccino',
-    price: '6oz',
-    type: 'Espresso'
-  },
-  {
-    id: 23,
-    itemName: 'Cortado ✧',
-    price: '4oz',
-    type: 'Espresso'
-  },
-  {
-    id: 24,
-    itemName: 'Macchiato',
-    price: '3oz',
-    type: 'Espresso'
-  },
-  {
-    id: 25,
-    itemName: 'Americano',
-    price: 'hot or iced - 8oz / 12oz',
-    type: 'Espresso'
-  },
-  {
-    id: 26,
-    itemName: 'Espresso shot',
-    price: 'single / double',
-    type: 'Espresso'
-  },
-  {
-    id: 31,
-    itemName: 'Matcha Latte ✧',
-    price: 'hot or iced - 8oz / 12oz',
-    type: 'Tea'
-  },
-  {
-    id: 32,
-    itemName: 'Various loose leaf tea',
-    price: '8oz / 12oz',
-    type: 'Tea'
-  },
-  {
-    id: 41,
-    itemName: 'Espresso Martini ✧',
-    price: 'vodka / tequila / mezcal / whiskey',
-    type: 'Boozy'
-  }
-];
-
-export default menu;
