--- src/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <title>☕️ by Tyler</title>
-    <link rel="icon" href="favicon.ico">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.15.4/css/all.css">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Playfair+Display&family=Source+Sans+Pro:wght@400;600&display=swap">
-    <link rel="stylesheet" href="style.css">
-  </head>
-
-  <body>
-    <div id="app"></div>
-    <script src="index.bundle.js"></script>
-    <noscript>
-      <p class="message info-message"><span class="fa fa-info-circle fa-lg fa-fw" aria-hidden="true"></span> This site uses JavaScript. Please enable JavaScript in your browser.</p>
-    </noscript>
-  </body>
-</html>
