--- src/index.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-import { BrowserRouter } from 'react-router-dom';
-import App from './App';
-import 'file-loader?name=[name].[ext]!./index.html';
-import 'file-loader?name=[name].[ext]!./favicon.ico';
-import 'normalize.css';
-import './stylesheets/style.scss';
-
-createRoot(document.getElementById('app')).render(
-  <BrowserRouter>
-    <App />
-  </BrowserRouter>
-);
