--- src/stylesheets/footer.scss
@@ -1,80 +1,0 @@
-footer {
-  background: $dark-gray;
-  bottom: 0;
-  color: change-color($black, $lightness: 100%);
-  left: 0;
-  padding: 0 20px;
-  position: absolute;
-  right: 0;
-  width: 100%;
-  a,
-  a:visited,
-  a:hover,
-  a:active,
-  a:focus {
-    color: $beige;
-  }
-  hr {
-    border-bottom: 1px solid $beige;
-  }
-  .coffee-shop-footer {
-    padding: 0 20px;
-    .site-links {
-      display: -webkit-box;
-      display: -ms-flexbox;
-      display: flex;
-      -webkit-box-pack: center;
-      -ms-flex-pack: center;
-      justify-content: center;
-      padding: 14px 0 26px;
-      a,
-      a:visited,
-      a:hover,
-      a:active,
-      a:focus {
-        color: change-color($black, $lightness: 100%);
-        font-size: 0.85rem;
-        font-weight: 600;
-        letter-spacing: 3px;
-        text-transform: uppercase;
-        &:not(:last-child) {
-          margin-right: 10px;
-        }
-        &:hover {
-          color: $light-gray;
-          text-decoration: none;
-        }
-      }
-    }
-    .other {
-      display: -webkit-box;
-      display: -ms-flexbox;
-      display: flex;
-      -webkit-box-align: center;
-      -ms-flex-align: center;
-      align-items: center;
-      -webkit-box-orient: vertical;
-      -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-      flex-direction: column;
-      @media all and (min-width: 620px) {
-        -webkit-box-orient: horizontal;
-        -webkit-box-direction: normal;
-        -ms-flex-direction: row;
-        flex-direction: row;
-        -webkit-box-pack: justify;
-        -ms-flex-pack: justify;
-        justify-content: space-between;
-      }
-      p {
-        font-size: 0.9rem;
-        margin: 6px 0;
-      }
-    }
-  }
-  .app-footer {
-    font-size: 0.8rem;
-    padding: 14px 0;
-    text-align: center;
-  }
-}
