--- src/stylesheets/header.scss
@@ -1,36 +1,0 @@
-header {
-  background:$dark-brown;
-  color: change-color($black, $lightness: 100%);
-  height: 100px;
-  left: 0;
-  position: fixed;
-  right: 0;
-  text-align: center;
-  top: 0;
-  width: 100%;
-  z-index: 1;
-  nav {
-    padding: 14px 0;
-    a,
-    a:visited,
-    a:hover,
-    a:active,
-    a:focus {
-      color: change-color($black, $lightness: 100%);
-      font-size: 0.85rem;
-      font-weight: 600;
-      letter-spacing: 3px;
-      text-transform: uppercase;
-      &:not(:last-child) {
-        margin-right: 10px;
-      }
-      &:hover {
-        color: $light-gray;
-        text-decoration: none;
-      }
-      &.active {
-        color: $beige;
-      }
-    }
-  }
-}
