--- src/stylesheets/hero-banner.scss
@@ -1,49 +1,0 @@
-.hero-banner {
-  background: $golden-brown;
-  position: relative;
-  .hero-banner-image {
-    background-position: center center;
-    background-repeat: no-repeat;
-    background-size: cover;
-    -webkit-filter: brightness(60%);
-    filter: brightness(60%);
-    min-height: 300px;
-    width: 100%;
-    @media all and (min-width: 600px) {
-      min-height: 400px;
-    }
-    @media all and (min-width: 800px) {
-      background-position: top center;
-      min-height: 500px;
-    }
-    @media all and (min-width: 900px) {
-      min-height: 600px;
-    }
-    @media all and (min-width: 1000px) {
-      min-height: 700px;
-    }
-    @media all and (min-width: 1100px) {
-      min-height: 800px;
-    }
-  }
-  .hero-banner-content {
-    color: $beige;
-    left: 0;
-    letter-spacing: 6px;
-    position: absolute;
-    top: 42%;
-    width: 100%;
-    h2 {
-      font-size: 2.5rem;
-      margin: 0;
-      text-align: center;
-      text-transform: uppercase;
-      @media all and (min-width: 600px) {
-        font-size: 3rem;
-      }
-      @media all and (min-width: 800px) {
-        font-size: 3.5rem;
-      }
-    }
-  }
-}
