--- src/stylesheets/main.scss
@@ -1,270 +1,0 @@
-main {
-  padding-top: 100px;
-  padding-bottom: 385px;
-  @media all and (min-width: 620px) {
-    padding-bottom: 330px;
-  }
-  .main-content {
-    margin: 0 auto;
-    max-width: 600px;
-    padding: 60px 20px 0;
-    hr {
-      border-bottom: 2.5px solid $dark-brown;
-      margin: 38px 12px;
-    }
-    .bold {
-      font-weight: 600;
-    }
-    .italics {
-      font-style: italic;
-    }
-  }
-}
-
-// Styles for pages with a hero banner
-.hero-banner + .main-content {
-  h3 {
-    border-bottom: 6px double $golden-brown;
-    color: $dark-brown;
-    font-size: 1.8rem;
-    letter-spacing: 3px;
-    margin: 0 auto;
-    padding: 0 15px;
-    text-align: center;
-    text-transform: uppercase;
-    width: -webkit-fit-content;
-    width: -moz-fit-content;
-    width: fit-content;
-    @media all and (min-width: 620px) {
-      font-size: 2.2rem;
-    }
-  }
-  h4 {
-    color: $dark-brown;
-    font-size: 1.4rem;
-    letter-spacing: 2px;
-    margin: 0;
-    @media all and (min-width: 620px) {
-      font-size: 1.6rem;
-    }
-  }
-}
-
-// Homepage styles
-.main-content.home {
-  max-width: 100%;
-  padding: 70px 0 20px;
-  .content-container.welcome {
-    font-size: 1.7rem;
-    margin: 0 auto;
-    max-width: 1000px;
-    padding: 0 40px 80px;
-    text-align: center;
-    p {
-      margin: 0;
-    }
-  }
-  .content-container.image-gallery {
-    display: -ms-grid;
-    display: grid;
-    grid-gap: 20px;
-    grid-template-columns: 40% 40%;
-    -webkit-box-pack: center;
-    -ms-flex-pack: center;
-    justify-content: center;
-    padding-bottom: 80px;
-    @media all and (min-width: 620px) {
-      grid-template-columns: 30% 30% 30%;
-    }
-    @media all and (min-width: 820px) {
-      grid-template-columns: 20% 20% 20% 20%;
-    }
-  }
-  .content-container.info {
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-align: center;
-    -ms-flex-align: center;
-    align-items: center;
-    -webkit-box-orient: vertical;
-    -webkit-box-direction: normal;
-    -ms-flex-direction: column;
-    flex-direction: column;
-    margin: 0 auto;
-    max-width: 1000px;
-    @media all and (min-width: 820px) {
-      -webkit-box-orient: horizontal;
-      -webkit-box-direction: normal;
-      -ms-flex-direction: row;
-      flex-direction: row;
-    }
-    .info-container {
-      display: -webkit-box;
-      display: -ms-flexbox;
-      display: flex;
-      -webkit-box-orient: vertical;
-      -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-      flex-direction: column;
-      height: 350px;
-      -webkit-box-pack: center;
-      -ms-flex-pack: center;
-      justify-content: center;
-      margin: 0 20px;
-      padding: 0 20px;
-      text-align: center;
-      width: 90%;
-      @media all and (min-width: 820px) {
-        height: 400px;
-        width: 50%;
-      }
-      &:last-child {
-        margin-top: 20px;
-        @media all and (min-width: 820px) {
-          margin-top: 0;
-          margin-left: 0;
-        }
-      }
-      h2 {
-        font-size: 1.3rem;
-      }
-    }
-    .menu-info {
-      background: $dark-brown;
-      color: change-color($black, $lightness: 100%);
-      .button,
-      .button:visited,
-      .button:hover,
-      .button:active,
-      .button:focus {
-        background: change-color($black, $lightness: 100%);
-        border: 2px solid change-color($black, $lightness: 100%);
-        color: $dark-brown;
-        margin: 0 auto;
-        width: -webkit-fit-content;
-        width: -moz-fit-content;
-        width: fit-content;
-        &:hover {
-          background: transparent;
-          color: change-color($black, $lightness: 100%);
-        }
-      }
-    }
-    .wifi-info {
-      background: change-color($black, $lightness: 100%);
-      color: $dark-brown;
-      p {
-        font-size: 1.15rem;
-      }
-      .wifi-password {
-        font-size: 1.25rem;
-        font-weight: 600;
-      }
-    }
-  }
-}
-
-// About page styles
-.main-content.about {
-  max-width: 800px;
-  .content-container {
-    padding: 20px;
-    &:not(:last-child) {
-      margin-bottom: 40px;
-    }
-  }
-  h3 {
-    border-bottom: 6px double $golden-brown;
-  }
-  figure {
-    padding: 20px;
-  }
-  img {
-    margin: 0 auto;
-    max-width: 600px;
-    width: 100%;
-    -webkit-box-shadow: 1px 1px 8px 1px rgba(change-color($black, $lightness: 43%), 0.7);
-    box-shadow: 1px 1px 8px 1px rgba(change-color($black, $lightness: 43%), 0.7);
-    box-shadow: 1px 1px 8px 1px rgba(change-color($black, $lightness: 43%), 0.7);
-  }
-  p {
-    line-height: 1.35;
-  }
-}
-
-// Menu page styles
-.main-content.menu {
-  h3 {
-    margin-bottom: 30px;
-  }
-  h4 {
-    border-bottom: 2px solid $golden-brown;
-    margin: 38px 0 18px;
-    width: -webkit-fit-content;
-    width: -moz-fit-content;
-    width: fit-content;
-  }
-  .menu-item {
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-orient: vertical;
-    -webkit-box-direction: normal;
-    -ms-flex-direction: column;
-    flex-direction: column;
-    padding: 8px 0;
-    @media all and (min-width: 500px) {
-      -webkit-box-orient: horizontal;
-      -webkit-box-direction: normal;
-      -ms-flex-direction: row;
-      flex-direction: row;
-      -webkit-box-pack: justify;
-      -ms-flex-pack: justify;
-      justify-content: space-between;
-      padding: 6px 0;
-    }
-    .item-name {
-      font-size: 1.15rem;
-      font-weight: 600;
-
-    }
-    .price {
-      font-size: 0.85rem;
-    }
-  }
-  .menu-info {
-    font-size: 0.9rem;
-  }
-}
-
-// Visit page styles
-.main-content.visit {
-  h3:not(:first-child) {
-    margin-top: 40px;
-  }
-  p {
-    font-size: 1.1rem;
-    font-weight: 600;
-    text-align: center;
-    @media all and (min-width: 620px) {
-      font-size: 1.3rem;
-    }
-  }
-}
-
-// 404 page styles
-.main-content.not-found {
-  h2 {
-    color: $dark-brown;
-    font-size: 2.8rem;
-    line-height: 1.2;
-    text-align: center;
-    text-transform: uppercase;
-  }
-  .button-group {
-    -webkit-box-pack: center;
-    -ms-flex-pack: center;
-    justify-content: center;
-  }
-}
