--- src/stylesheets/style.scss
@@ -1,211 +1,0 @@
-@font-face {
-  font-family: "Dark Idea";
-  src: url(.././assets/fonts/dark-idea.ttf);
-}
-
-@font-face {
-  font-family: "TA Typefire";
-  src: url(.././assets/fonts/ta-typefire.otf);
-}
-
-@import "variables";
-
-// BASE PAGE STYLES
-
-* {
-  -webkit-box-sizing: border-box;
-  box-sizing: border-box;
-}
-
-html {
-  background: transparent;
-  font-family: $main-font;
-  height: 100%;
-  letter-spacing: 1px;
-  body {
-    background: $beige;
-    color: $black;
-    min-height: 100%;
-    -ms-scroll-chaining: none;
-    overscroll-behavior: none;
-    padding: 0;
-    position: relative;
-    h1,
-    h2,
-    h3,
-    h4,
-    h5,
-    h6 {
-      font-family: $heading-font;
-      margin-top: 0;
-    }
-    h1 {
-      font-weight: 400;
-      margin: 0;
-    }
-  }
-}
-
-a,
-a:visited,
-a:hover,
-a:active,
-a:focus {
-  color: $dark-brown;
-  text-decoration: none;
-  &:hover {
-    text-decoration: underline;
-  }
-}
-
-a:focus,
-button:focus,
-details:focus,
-input:focus,
-select:focus,
-textarea:focus,
-[tabindex]:focus {
-  -webkit-box-shadow: 0 0 3px 3px hsla(203, 95%, 62%, 0.8);
-  box-shadow: 0 0 3px 3px hsla(203, 95%, 62%, 0.8);
-  outline: none;
-}
-
-img {
-  display: block;
-  height: auto;
-  max-width: 100%;
-}
-
-figure {
-  margin: 0;
-}
-
-hr {
-  border: 0;
-}
-
-// Button styles
-.button-group {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-}
-
-.button,
-.button:visited,
-.button:hover,
-.button:active,
-.button:focus {
-  background: $dark-brown;
-  border: 2px solid $dark-brown;
-  border-radius: 28px;
-  color: $beige;
-  color: change-color($black, $lightness: 100%);
-  cursor: pointer;
-  font-size: 0.85rem;
-  font-weight: 600;
-  letter-spacing: 1px;
-  padding: 12px 16px;
-  text-transform: uppercase;
-  &:hover {
-    background: transparent;
-    color: $dark-brown;
-    text-decoration: none;
-  }
-}
-
-// Logo styles
-.logo {
-  font-size: 1.8rem;
-  margin: 0 auto;
-  padding-top: 14px;
-  text-align: center;
-  width: -webkit-fit-content;
-  width: -moz-fit-content;
-  width: fit-content;
-  a,
-  a:visited,
-  a:hover,
-  a:active,
-  a:focus {
-    color: $beige;
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-orient: vertical;
-    -webkit-box-direction: normal;
-    -ms-flex-direction: column;
-    flex-direction: column;
-    &:hover {
-      text-decoration: none;
-    }
-    .logo--top {
-      font-family: $logo-font--coffee;
-      margin-bottom: -10px;
-    }
-    .logo--bottom {
-      font-family: $logo-font--grounds;
-    }
-  }
-}
-
-.social-media-links {
-  margin-top: 14px;
-  padding-bottom: 26px;
-  -webkit-box-ordinal-group: 0;
-  -ms-flex-order: -1;
-  order: -1;
-  @media all and (min-width: 620px) {
-    padding-bottom: 0;
-    -webkit-box-ordinal-group: 1;
-    -ms-flex-order: 0;
-    order: 0;
-  }
-  a,
-  a:visited,
-  a:hover,
-  a:active,
-  a:focus {
-    color: change-color($black, $lightness: 100%);
-    font-size: 1.5rem;
-    &:hover {
-      color: $beige;
-    }
-    &:not(:last-child) {
-      margin-right: 6px;
-    }
-  }
-}
-
-// Message styles
-.message {
-  border-radius: 5px;
-  line-height: 1.55;
-  margin: 20px auto;
-  max-width: 800px;
-  padding: 15px 10px;
-  text-align: center;
-}
-
-.info-message {
-  background: change-color($info-blue, $lightness: 95%);
-  border: 1px solid change-color($info-blue, $lightness: 90%);
-  color: $info-blue;
-}
-
-noscript .info-message {
-  left: 50%;
-  margin: 0 auto;
-  min-width: 280px;
-  position: absolute;
-  text-align: center;
-  top: 50%;
-  -webkit-transform: translate(-50%, -50%);
-  -ms-transform: translate(-50%, -50%);
-  transform: translate(-50%, -50%);
-}
-
-@import "header";
-@import "main";
-@import "footer";
-@import "hero-banner";
