--- src/stylesheets/variables.scss
@@ -1,12 +1,0 @@
-$black: hsl(0, 0%, 0%);
-$dark-gray: hsl(75, 3%, 25%);
-$light-gray: hsl(196, 9%, 80%);
-$beige: #e4cfb4;
-$golden-brown: #9a5928;
-$dark-brown: #432d2d;
-$info-blue: hsl(204, 100%, 26%);
-
-$main-font: "Source Sans Pro", sans-serif;
-$heading-font: "Playfair Display", serif;
-$logo-font--coffee: "Dark Idea", "Source Sans Pro", sans-serif;
-$logo-font--grounds: "TA Typefire", "Playfair Display", serif;
