--- webpack.config.js
@@ -1,46 +1,0 @@
-const path = require('path');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-
-module.exports = {
-  output: {
-    path: path.join(__dirname, 'public'),
-    filename: 'index.bundle.js',
-    assetModuleFilename: 'assets/[hash][ext][query]'
-  },
-  devtool: 'eval-cheap-source-map',
-  devServer: {
-    port: 8080,
-    static: [{
-      watch: true
-    }],
-    historyApiFallback: true
-  },
-  module: {
-    rules: [
-      {
-        test: /\.(js|jsx)$/,
-        exclude: /node_modules/,
-        use: {
-          loader: 'babel-loader'
-        }
-      },
-      {
-        test: /\.(sc|c)ss$/,
-        use: [
-          MiniCssExtractPlugin.loader,
-          'css-loader',
-          'sass-loader'
-        ]
-      },
-      {
-        test: /\.(jpg|png|svg|ttf|otf|woff)$/,
-        type: 'asset'
-      }
-    ]
-  },
-  plugins: [
-    new MiniCssExtractPlugin({
-      filename: 'style.css'
-    })
-  ]
-};
